add event handling for token create, update, delete
