+++ packages/manager/src/App.tsx
@@ -34,6 +34,7 @@ import GoTo from './GoTo';
 import { databaseEventsHandler } from './queries/databases';
 import { domainEventsHandler } from './queries/domains';
 import { volumeEventsHandler } from './queries/volumes';
+import { tokenEventHandler } from './queries/tokens';
 
 interface Props {
   toggleTheme: () => void;
@@ -125,6 +126,13 @@ export class App extends React.Component<CombinedProps, State> {
     events$
       .filter((event) => event.action.startsWith('volume') && !event._initial)
       .subscribe(volumeEventsHandler);
+
+    /*
+     * Send any Token events to the Token events handler in the queries file
+     */
+    events$
+      .filter((event) => event.action.startsWith('token') && !event._initial)
+      .subscribe(tokenEventHandler);
 
     /*
      * We want to listen for migration events side-wide
